@@ -5,22 +5,23 @@
     "keywords": ["laravel", "framework"],
     "license": "MIT",
     "require": {
-        "php": "^8.1",
-        "guzzlehttp/guzzle": "^7.2",
-        "laravel/breeze": "^1.27",
-        "laravel/framework": "^10.10",
-        "laravel/sanctum": "^3.3",
-        "laravel/tinker": "^2.8"
+        "php": "^8.3",
+        "guzzlehttp/guzzle": "^7.8.1",
+        "juststeveking/sdk-tools": "^0.0.5",
+        "laravel/framework": "^10.38.2",
+        "laravel/sanctum": "^3.3.2",
+        "laravel/tinker": "^2.8.2"
     },
     "require-dev": {
-        "fakerphp/faker": "^1.9.1",
-        "laravel/pint": "^1.0",
-        "laravel/sail": "^1.18",
-        "mockery/mockery": "^1.4.4",
-        "nunomaduro/collision": "^7.0",
-        "pestphp/pest": "^2.0",
-        "pestphp/pest-plugin-laravel": "^2.0",
-        "spatie/laravel-ignition": "^2.0"
+        "fakerphp/faker": "^1.23.0",
+        "laravel/pint": "^1.13.7",
+        "laravel/sail": "^1.26.3",
+        "mockery/mockery": "^1.6.7",
+        "nunomaduro/collision": "^7.10",
+        "pestphp/pest": "^2.28.1",
+        "pestphp/pest-plugin-laravel": "^2.2",
+        "roave/security-advisories": "dev-latest",
+        "spatie/laravel-ignition": "^2.3.3"
     },
     "autoload": {
         "psr-4": {
