@@ -1,17 +1,17 @@
 <?php
 
+declare(strict_types=1);
+
 namespace App\Http\Middleware;
 
 use Illuminate\Routing\Middleware\ValidateSignature as Middleware;
 
-class ValidateSignature extends Middleware
+final class ValidateSignature extends Middleware
 {
     /**
-     * The names of the query string parameters that should be ignored.
-     *
-     * @var array<int, string>
+     * @var array<int,string>
      */
-    protected $except = [
+    protected array $except = [
         // 'fbclid',
         // 'utm_campaign',
         // 'utm_content',
