@@ -1,15 +1,15 @@
 <?php
 
+declare(strict_types=1);
+
 namespace App\Http\Middleware;
 
 use Illuminate\Foundation\Http\Middleware\TrimStrings as Middleware;
 
-class TrimStrings extends Middleware
+final class TrimStrings extends Middleware
 {
     /**
-     * The names of the attributes that should not be trimmed.
-     *
-     * @var array<int, string>
+     * @var array<int,string>
      */
     protected $except = [
         'current_password',
