@@ -1,15 +1,15 @@
 <?php
 
+declare(strict_types=1);
+
 namespace App\Http\Middleware;
 
 use Illuminate\Http\Middleware\TrustHosts as Middleware;
 
-class TrustHosts extends Middleware
+final class TrustHosts extends Middleware
 {
     /**
-     * Get the host patterns that should be trusted.
-     *
-     * @return array<int, string|null>
+     * @return array<int,string|null>
      */
     public function hosts(): array
     {
