@@ -7,13 +7,10 @@ use Illuminate\Foundation\Application;
 
 trait CreatesApplication
 {
-    /**
-     * Creates the application.
-     */
+
     public function createApplication(): Application
     {
         $app = require __DIR__.'/../bootstrap/app.php';
-
         $app->make(Kernel::class)->bootstrap();
 
         return $app;
